requirements.txt
asgiref==3.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
Django==3.2.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
gunicorn==20.1.0
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.3.2
Pillow==8.4.0
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
whitenoise==5.3.0
zipp==3.6.0
